Adding static function to the schema class
